Optimize WTF::Vector::Contains().
This optimizes WTF::Vector::Contains() by copying the code used by Find
rather than re-using it, so that the compiler doesn't need to generate
extra code to handle the possibility that the kNotFound-th element in
the array was the one that was found.

This doesn't seem to make a measurable difference on Speedometer3,
although I did find it while looking at the assembly for a hot function
in one of the tests.  Having simpler generated code seems like a win,
though.

Change-Id: I575731027eb6ce67e570af0a9dab382289ec1750
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5493389
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Auto-Submit: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293667}
NOKEYCHECK=True
GitOrigin-RevId: d791649bc979aa4fead7e8b49948eb1b24f69184