Disallow atomic moves when moving across shadow/document
Using the node's TreeScope as a way to ensure this is valid.
Added new tests that rely on the CSS transition behavior of moveBefore.

This narrows down the impact surface of moveBefore, preventing potential
unknown edge cases. We can revise this and enable same-document
cross-shadow atomic moves in the future if the use case arises and the
implications are clear.

Bug: 40150299
Change-Id: Ie91afb678ee1b97d68d76337c0bbc5e04b6accdb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5465565
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293836}
NOKEYCHECK=True
GitOrigin-RevId: fe970cf8664036737ad5be50e849e7e886fb09b5