Moving zoom from WebViewImpl to WebFrameWidgetImpl
As we allow CSS zoom to affect iframes, zoom can no longer be a property
of a web view, as a web view can contain several frames, each with a
different zoom. This change removes the zoom_level_ variable and
the functions associated with it from WebViewImpl,and moves it to the
frame level.

Bug: 329482480

Change-Id: I1136e3a40f0c07587b584055ac4348fc9c9b2b2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5402103
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Auto-Submit: Yotam Hacohen <yotha@chromium.org>
Commit-Queue: Yotam Hacohen <yotha@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293975}
NOKEYCHECK=True
GitOrigin-RevId: 3d6aca656f25ee76fa641edd9882aa87bdcb1e7e