Fixed canvas 2d layer when a composited global shadow is used
According to the WHATWG spec for global shadows, the shadow and
foreground have to be composited independently one after the other
against the destination surface. In other word, the foreground
gets composited on top of the shadow.
https://html.spec.whatwg.org/multipage/canvas.html#drawing-model

This was broken for layers. The way we implement shadows for normal
draw calls is essentially by drawing the shape twice: once for the
shadow and once for the foreground. Doing this for layers however
would require drawing the whole layer twice, which would be
expensive. We instead have to get a hold on the layer's resulting
texture and draw that texture twice.

This can be done using a feature recently added to Skia allowing
layers to have a FilterSpan instead of a single filter. Each filter
in the FilterSpan is applied in sequence on the layer's output.
https://skia-review.googlesource.com/c/skia/+/784549

By using a FilterSpan comprised of a shadow filter and an identity
(nullptr) filter, we get the effect of compositing a shadow and then
the foreground.

On the Skia side, the FilterSpan is a property of the layer and is
not part of the SkPaint object, where normal filters are. To align
with this, this CL doesn't add the FilterSpan into cc::PaintFlags.
Doing so would add the overhead of carrying a vector of filters for
all paint ops, when it's only needed and supported for layers.
Instead, a new "SaveLayerFiltersOp" is added, isolating the
overhead to this one particular paint op.

Fixed: 40257559
Change-Id: Ie1c29a25f28efadabaf01c589837e4295a29d934
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5451004
Reviewed-by: Rick Byers <rbyers@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Brian Osman <brianosman@google.com>
Cr-Commit-Position: refs/heads/main@{#1294597}
NOKEYCHECK=True
GitOrigin-RevId: c8f84a07292e56314b57d734e7c2df1a9eb55047