Stop using SetHeaderPolicy in CreateFromParsedPolicy.
This also introduces a struct to carry the final result of parsing allowlist and reporting endpoints in the permissions policy header and we pass that struct to the PermissionsPolicy constructor.

This also make origin_ and inherited_policies_ const.

This leaves several places still passing {} and modifying with setters later.

Change-Id: Ia501de3b66d09e296fc35e7259d4200c5fed5c78
Bug: 338183241
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5475204
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1294828}
NOKEYCHECK=True
GitOrigin-RevId: 77bbdcf9186b009222701f8998491ad5f86a7638