Enable background shorthand animation to run on the compositor
Previously, the background-shorthand would fail to animate on the
compositor base KeyframeEffectModelBase::RequiresPropertyNode() was
iterating over all properties and not just dynamic properties.

Compositor animation tests did not have an image generator for
background color, and the eligibility check would fail as a result.
Added a fake image generator for testing the background shorthand.

Finally, the set of properties for an animation was being constructed
multiple times during the process of checking and then starting an
animation on the compositor. This property set is now cached.  The
cache is lazy constructed and reset whenever the keyframe groups is
reset.

Bug: 336596362, 337858293
Change-Id: Ieec142b67edf5bce747c60b7f8529b4841d3e77e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5491477
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1294910}
NOKEYCHECK=True
GitOrigin-RevId: a115ef6005618c3ba92de11c4c61994320a22fc6