Migrate tfhub_models to gcs first class
This hook is being migrated to be a GCS first class dep. This means
that the sha file is no longer needed. Any updates to this dep will
need to update the gcs dep entry in DEPS, instead of the sha file.

Bug: b/337331070
Change-Id: I7be49b9f0c79557f84ea5db478d2fc6793f9bb1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5499517
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Commit-Queue: Stephanie Kim <kimstephanie@google.com>
Cr-Commit-Position: refs/heads/main@{#1294987}
NOKEYCHECK=True
GitOrigin-RevId: 1b5dd66c49f7b6712a0bfb7cc59d72fd0dd892f0