Remove use_perfetto_client_library build flag in /third_party/blink/renderer/platform/widget/input
This flag was introduced to provide a smooth transition from legacy
tracing macros to Perfetto tracing SDK. Now the transition is complete,
and the flag is unconditionally set to true on all platforms. So we can
remove the flag and the code that is disabled by it.

This CL was uploaded by git cl split.

R=skobes@chromium.org

Bug: 42050015, 336718643
Change-Id: I3e07b6f7dcace379e38ce759d592c65e15f84311
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5506055
Commit-Queue: Steve Kobes <skobes@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Auto-Submit: Mikhail Khokhlov <khokhlov@google.com>
Cr-Commit-Position: refs/heads/main@{#1295055}
NOKEYCHECK=True
GitOrigin-RevId: 0f91799ce3b69bd483a2e7a871cbfa676ea79eac