Fix Release build failures in ChromiumVirtualDisplayDriver project
I tried build a release version of the driver and it failed because
the C++ version was not specified. I added that attribute and then
the build failed because the the properties header could not be found.
So I added that directory in the AdditionalIncludeDirectories
attribute and now I can build a Release version of the driver.

Bug: b:188675483
Change-Id: Ib689029b5b042d8ee3a8169d6e0f4744e6dcb93f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5506896
Reviewed-by: Brad Triebwasser <btriebw@chromium.org>
Commit-Queue: Brad Triebwasser <btriebw@chromium.org>
Auto-Submit: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1295124}
NOKEYCHECK=True
GitOrigin-RevId: 0b269e40637ac053e00034947a3779cc2cce2fa8