[Blink] Move unaccelerated resource prep into CanvasResource subclass
CanvasResource::PrepareUnacceleratedTransferableResource() will be
called only if the resource in question is a CanvasResourceSharedBitmap,
as (a) it is called only if SupportsAcceleratedCompositing() returns
false [1] and (b) CanvasResourceSharedBitmap is the only subclass that
returns false for SupportsAcceleratedCompositing() [2].

This CL makes PrepareUnacceleratedTransferableResource() virtual and
moves the current implementation into CanvasResourceSharedBitmap. This
paves the way for a followup that changes the implementation to operate
directly on the SharedBitmapId and invoke
TransferableResource::MakeSoftwareSharedBitmapId().

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/canvas_resource.cc;l=202-204;drc=6f3f85b321146cfc0f9eb81a74c7c2257821461e;bpv=1;bpt=1
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/canvas_resource.h;l=102;drc=6f3f85b321146cfc0f9eb81a74c7c2257821461e;bpv=1;bpt=1

Change-Id: I00a1ebcd836168fd0c7e79ace8d687478fa09779
Bug: 337538024
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5508597
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1295439}
NOKEYCHECK=True
GitOrigin-RevId: 7ad8360fda5cec49b8c3d99030ea10eb099f835e