Add content to appearance:base-select fallback button
This patch adds extra content and styling to the fallback button in the
UA shadowroot of <select> which gets rendered when it is
appearance:base-select and the author does not provide a <button>
element to replace it with.

In order to do this, this patch also adds 3 pseudo-elements to the
<select> element:
- ::select-fallback-button which targets the fallback <button>
- ::select-fallback-button-icon which targets the <div> inside the
  fallback <button> which renders an svg as its background-image
- ::select-fallback-selectedoption which targets the <span> inside the
  fallback <button> which gets the selected <option>'s text content
  copied into it.

The CSS and contents are all exactly the same as they are in
<selectlist>. <selectlist> also has the same pseudo elements.

This patch also changes some tests to inherit from PageTestBase. Adding
the svg background-image caused a scheduling crash in v8 with something
about microtask queues. I don't know what PageTestBase does differently
from these tests which are making their own TaskEnvironment, but
PageTestBase does it in a way which doesn't make the test crash.

Bug: 1511354
Change-Id: Iaf362dd1253f4dd9e5984cc3301c31e9e9ae558c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5464171
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1295509}
NOKEYCHECK=True
GitOrigin-RevId: f2af1cc58f57236069077da2f8986d274b3e76f3