Fix memory leak via TextFragmentFinder
TextFragmentFinder holds a Member to the Document over which it iterates
so it will keep its Document alive. However, The Finder is isn't cleared
when the search finishes or TextFragmentSelectorGenerator reset and so
is kept live via this ownership chain:

LocalFrame
  -> TextFragmentHandler
  -> TextFragmentSelectorGenerator
  -> TextFragmentFinder

The handler and selector generator is reset whenever a new document is
attached (from Document::Shutdown and LocalFrame::DetachImpl) but the
member isn't cleared so, in navigations that reuse a LocalFrame this can
keep the old document alive.

Once the finder operation completes we no longer need to keep it around so the fix here is to clear the TextFragmentFinder Member in the
generator when that happens.

Bug: 336357543
Change-Id: Id362ed7655ea781f37578d2cbdeffd71a68b213b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5499600
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1295575}
NOKEYCHECK=True
GitOrigin-RevId: 2a2797bdce6c8656038ac41894df64b815d3f8e5