view-transition: Invalidate paint when throttling local sub-frames.
ViewTransition relies on local sub-frame throttling to suppress
rendering opportunities. This throttling needs to be accompanied with
invalidating paint on the subframe, which is not being done currently.

This change adds this paint invalidating whenever throttling changes as
a result of ViewTransition.

Bug: 330917526
Change-Id: I4db695bc7e119e5fc795db011ae5900574f9220b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5499477
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1295623}
NOKEYCHECK=True
GitOrigin-RevId: 8330a8cd35e83f1e325decbafd5b1a00cc88eb90