Add crabbyavif placeholder wrappers
The files are exact copy of the libavif wrapper. This CL simply
adds the new files so that it easier to view the diff when the
actual crabbyavif wrapper is added in a follow-up.

This CL also does not add the new source files to BUILD.gn

Bug: b/323754052
Change-Id: If625963a34d20a65b486ab667ae0dbdd36ad6866
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5512826
Reviewed-by: James Zern <jzern@google.com>
Auto-Submit: Vignesh Venkat <vigneshv@google.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1295827}
NOKEYCHECK=True
GitOrigin-RevId: 49b14c6fe3c1a4d59a57bd7ee56674412becc157