Reduce memory usage of preloading ML model
If a user leaves their mouse hovering over a link, the model
will be repeatedly executed to produce new predictions. Each time,
we add another ExperimentalPreloadingPrediction which involves making
a copy of the predicted URL. This can cause large increases in
browser process memory usage.

From local experiments with the model, I haven't found a case where
the predicted probability of a click increases while dwelling on a
link. So executing the model repeatedly for a single mouse over is
of questionable value. We now only execute once per mouse over (while
leaving this controllable via finch param).

For the predictions we do still make, we avoid copying the predicted
URL for the match predicate. We instead only bind a hash of the URL
for comparison. We also reduce the size of
ExperimentalPreloadingPrediction.

While we're here, we have PreloadingDataImpl store the predictions
without the indirection through unique_ptr.

Bug: 338200075, 40278151
Change-Id: I9e4c649f14cc953f15ed7ba5c4034f7646625265
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5507055
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Iman Saboori <isaboori@google.com>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1295838}
NOKEYCHECK=True
GitOrigin-RevId: bf9af616d2ee258987429c8bb6aee5cb5b332944