Set RTCEncodedVideoFrame timestamp to capture_begin_time if available
This CL also modifies the test to start using getUserMedia as the
timeouts occur when we use getNoiseStream, which creates first frame
with timestamp 0 which leads to frame drops. We also add another test
that inserts a new VideoFrame with new timestamp.

We also modify the TestExpectations of the corresponding test to
FAIL/PASS/Timeout as this change should convert the test results
timeouts to fail/pass.

Doc: go/capture-timestamp-flake

Bug: 333739617, 338388034
Change-Id: I043009f05e21894a3bf8751c856bd2b6b646fbad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5509354
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Palak Agarwal <agpalak@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1296025}
NOKEYCHECK=True
GitOrigin-RevId: 6dd559335a5990c3c149256fbdce3ff0a600f6f0