view-transition: Introduce an effect node to group all subframe content
Add an effect node which can be the common ancestor for the
ViewTransitionEffect (the node that produces the root snapshot) and
the `::view-transition` pseudo-element. Since `::view-transition` is
after the isolation EffectNode for subframes, a change to add/remove the
ViewTransitionSubframeRootEffect needs to be isolation piercing so its
propagated to the `::view-transition`'s effect node.

Bug: 330917526
Change-Id: Id5565ad6e702c068b640c448905ea13cc43e7ab3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5513163
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Auto-Submit: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1296153}
NOKEYCHECK=True
GitOrigin-RevId: 54bb65501d6407a10a677ed819510bf1ebdc34dd