[Interop] Implement multi-arg CSS content alt text supporting strings and attr()
Alt text for CSS `content` was implemented a while ago [1], but while
the spec [2] allows an arbitrary number of arguments that can be
string, `attr()`, or counters, the implementation only supports a
single string argument.

This came to my attention as part of the Interop2024 [3] accessibility
focus area investigation, where some of the accessible name
computation subtests [4] fail due to the lack of full support for this
feature.

In this CL, expand support to allow an arbitrary number of arguments
and allow the arguments to be `attr()` as well as strings. Counter
support is more complicated since counter values are calculated
as part of layout tree computation, and the alt text does not
participate in the layout tree. So, counter support is left to a
future CL.

This CL is enough to get all current WPTs for the feature (that
I saw) passing. Interestingly, though WebKit implements the
multi-arg syntax and allows `attr()`, they also do not support
counters for this.

[1] https://www.chromestatus.com/features/4550056227110912
[2] https://drafts.csswg.org/css-content/#valdef-content---string--counter--attr
[3] https://wpt.fyi/interop-2024
[4] https://wpt.fyi/results/accname/name/comp_name_from_content.html?label=master&label=experimental&aligned&view=interop&q=label%3Ainterop-2024-accessibility

Change-Id: I7a6422d26b9857f1c54509349e7f582d43fbf1d1
Bug: 325612090, 338392169
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5502620
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1296164}
NOKEYCHECK=True
GitOrigin-RevId: 53e22c498667128d8841b082894cd59bcba51b7a