Use AbsoluteLocation() for DelegatedInkMetadata
This CL changes how a delegated ink metadata is created in Blink.
It replaces the use of the pointer event's x and y coordinates with
the pointer event's absolute location. This eliminates the need for
LayoutUnits in the calculations, therefore aligning it more closely
with the DelegatedInkPoint coordinate computation in the browser
process.

Change-Id: I2a3767a51a30980c94206612093cb71573415102
Bug: 338094916
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5499758
Commit-Queue: Sahir Vellani <sahir.vellani@microsoft.com>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1296203}
NOKEYCHECK=True
GitOrigin-RevId: 4875d63b15d3577f19a4e9bbe22ff40dfee31de4