blink: potentially avoid calling synchronizeattribute
SynchronizeAttribute is really only needed for a couple of things:
kStyleAttr or any svg-animatable property. A cheap way to check for
these two is if the qualifedname is defined. This catches way more
than is necessary, but good enough for a fast check.

Change-Id: I3d8a8538b6c4852171fe738bfed80a47c1a2e0f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5500948
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1296863}
NOKEYCHECK=True
GitOrigin-RevId: 1ce0e9e16644e34e8fb5a982fab775eced4cbdef