Make ShadowRoot's serializable property readonly
Spec is here:

  https://dom.spec.whatwg.org/#interface-shadowroot

This was an oversight in the initial code. Due to the fact that
`serializable` only recently shipped, and this is a corner case
usage pattern (setting serializable explicitly on a ShadowRoot),
I'd like to treat this as a bug fix. I think the risk is
extremely small, and this likely won't be noticed.

Bug: 40775629
Fixed: 336352736
Change-Id: I2715ffd9652bcfb71a20b5a0cd9882d51a0f88d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5516362
Auto-Submit: Mason Freed <masonf@chromium.org>
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1297008}
NOKEYCHECK=True
GitOrigin-RevId: 0a2971928054990f0485f5afaaf66dc397ef1d33