Move value parser end-of-stream handling into the caller.
The caller already checked end-of-stream itself for longhands,
but now also does so for shorthands. Importantly, all longhands
and shorthands are now changed so that they do not return error
on “not at end-of-stream”. This is in preparation for moving to
not stripping !important before calling the custom parsers,
which is in turn in preparation for moving to the streaming parser
for value parsing.

After this, any calls in longhands_custom.cc or shorthands_custom.cc
to range.AtEnd() is to be viewed with suspicion. They are not _always_
an error, but they need to be examined carefully.

Change-Id: I41bcd8ee37d77682e0560acc96f5074bd1367e24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5509593
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1297410}
NOKEYCHECK=True
GitOrigin-RevId: 169cc6716e6e869ac8bd94b725bafe1595e84311