Prerender: Cancel prerendering on add/removeJavascriptInterface()
When a WebView host app calls addJavascriptInterface(), the app expects
the JS interfaces are available on pages newly loaded after that.
However, if this API is called when prerendered pages exist, the JS
interfaces are not injected to the existing prerendered pages but
activation of these pages looks like new page loads to the app. This
could confuse the host app. To avoid the situation, This CL cancels all
existing prerendered pages when the API is called.

For the same reason, this CL also cancels prerendered pages on
removeJavascriptInterface().

NO_IFTTT=Changes for PrerenderFinalStatus will be done in separate repositories.

Change-Id: I2360c375759416d2c663afeed894b6fa8489957e
Bug: 331250164
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5500183
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1297850}
NOKEYCHECK=True
GitOrigin-RevId: c1e580167751bb6ea748b87d3605522a35c18b87