Update MediaVideoVisibilityTracker to observe fullscreen change events
This change updates the MediaVideoVisibilityTracker to detach whenever
the video, or any other element within the page, enters fullscreen.

The reasoning behind this change is that once the user has entered
fullscreen, we can assume that entering picture in picture is not
desired.

As an added benefit, this may also reduce the total number of nodes
we hit test.

Bug: 40275580
Change-Id: I64f8f7e86aac354e0906d606386ea697d018659b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5507016
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Benjamin Keen <bkeen@google.com>
Cr-Commit-Position: refs/heads/main@{#1297913}
NOKEYCHECK=True
GitOrigin-RevId: c5f5203ac537f27e5cc5f6b103345aa18702f371