[PWA] Implement PWA.launch
The change itself seems very straightforward except for the DCHECK-
failure caused by out-of-scope url. So the PWAHandler needs to check
the url scope manually to avoid crashing.
Due to the potential race-condition, the check is not 100% safe,
especially in the DCHECK_IS_ON(). But the DCHECK in question will be
removed sooner or later (http://crbug.com/338406726), so it's unlikely
a blocking issue for now.

Other than that, there are some less interesting changes to keep the
files in a good shape, e.g. the callback always goes to the end of
the parameter list, input parameters are always const references for
an easy global optimization from the base::BindOnce if any.

Bug: 331214986, 338406726
Change-Id: I81ecfe48a33dc7c6ad9f8571005570eeaf1f7597
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5503419
Commit-Queue: Zijie He <zijiehe@google.com>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1297964}
NOKEYCHECK=True
GitOrigin-RevId: 5474558a2f842b5f29c15874eefb75cd17f4f2c9