[v8][fastcall] Remove fallback for fast API users
So far, target functions of fast API calls were not allowed to
throw exceptions. Instead, a fallback mechanism was in place where
the target function of the fast API call would return to V8 with a
signal that it wanted to throw an exception, and V8 would call a
generic version of the same function to actually throw the
exception.

V8 recently added the capability to fast API calls to handle
exceptions. Therefore, the target function of a fast API call can
now throw an exception directly. Additionally, it is allowed now to
allocate objects, trigger a GC, and call back to JavaScript.

This CL changed API functions and the binding code that used the
fallback mechanism. For that the bindings generator is adjusted.

The document [1] contains the changes to the generated bindings
code of the WebGL function `uniform4fv(WebGLUniformLocation? location,
sequence<GLfloat> v)`.

[1] http://doc/1JD_k10zx4fNQSbKbkv_zW0-BgePOSOzgE1FvMiirmQU

Bug: 41492790
Change-Id: I2b26e3ffa74b134352ecf8c5552982dc5f163fe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5471803
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1298038}
NOKEYCHECK=True
GitOrigin-RevId: d9bfd74484c34eeed486e4d47d12ab6f2498b5c7