Strip !important after parsing values, not before.
This is a preparation for parsing CSS values using the streaming parser,
where we have no choice but to do so.

Note that for unparsed values (values that contain variable references),
we still do the !important stripping before. This will need to change
at some point in the future.

Change-Id: I44b1cde6b18f2fea61940aed36e17b932e70ed81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5509594
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1298059}
NOKEYCHECK=True
GitOrigin-RevId: cac78afc516ec6fdaab5d891c4d18a216ec45aa1