Fix UAF issue in BackgroundProcessor::OnDataPipeReadable
There is an UAF issue in BackgroundProcessor::OnDataPipeReadable.
BackgroundProcessor::MaybeStartProcessingResponse() is calling
SimpleWatcher::Watch() with a bound WTF::Unretained(this). This is not
safe. BackgroundProcessor::OnDataPipeReadable() can be called even after
the BackgroundProcessor is deleted in the main thread (not in the
background thread). This is because the SimpleWatcher will not be
canceled until the deletion task is executed in the background thread.

To fix this issue, this CL changes BackgroundProcessor::
MaybeStartProcessingResponse() method to call SimpleWatcher::Watch()
with a scoped_refptr(this), and changes BackgroundProcessor::Cancel()
to reset the SimpleWatcher.

Fixed: 337998760
Change-Id: I7d376a25dad3a992510e3575075b0e471ef4ebf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5514715
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1298079}
NOKEYCHECK=True
GitOrigin-RevId: 125bc950b30728d400f1dadc46715443f6d011e2