blink: conditionally create UniqueElementData when moving documents
When an element moves between documents and the quirks mode of the
documents differ, the AttributeData is made unique. This
is done because the id/class in quirks mode are lower case. This
patch makes this conditional based on whether the id/class is
lower-case. Specifically, if the id/class is already lower-case,
then there isn't a behavior change between the two modes and
no need to make ElementData unique.

Bug: None
Change-Id: I97daa2e5f90dafa3ac2a10a693ebfc516f67ebd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5519296
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1298342}
NOKEYCHECK=True
GitOrigin-RevId: c6ef27c04278400a057bea86723858b495e39874