Set ResourceStatus for SimplifyLoadingTransparentPlaceholderImage
Previously, when the SimplifyLoadingTransparentPlaceholderImage flag was
enabled, we created a transparent placeholder bitmap image in
ImageLoader::DoUpdateFromElement. However, we never set the content
status on the image, so it remained as "not started". This caused the
image onload event to not fire.

By using ImageResourceContent::CreateLoaded, we ensure that the image
content status is set to "cached", and hence loaded successfully, so
that the image onload event fires.

Bug: 336681952
Change-Id: Ibf05a59b21690ef265c1788ae2b32da5d4715445
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5520481
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1298462}
NOKEYCHECK=True
GitOrigin-RevId: 417b4749e78dab1c7cdae2a7cd533ba5fb624397