Make browser_preferred_color_scheme light when custom theme is on.
Issue:
When chrome has a browser theme on that changes the browser's color and
the user preferences are set to dark mode, UsedColorSchemeRootScrollbars
makes root scrollbars be dark in pages where no color scheme is
specified. This becomes an issue when the browser's theme is a light
color, which contrasts too starkly with the dark scrollbars. To repro
this issue, please see the attached bug.

The `browser_preferred_color_scheme` flag is used to send the preferred
color scheme from the browser process to blink in the renderer process
to help with the selection of a color scheme for scrollbars.

Fix:
- This CL changes the `browser_preferred_color_scheme` browser setting
to be set to the default value (kLight color scheme) when a custom theme
is detected in the browser. With this change,
UsedColorSchemeRootScrollbars doesn't override the root scrollbar by
changing it to dark mode.
- The `browser_preferred_color_scheme` setting is also being renamed to
better represent what it pertains to, and is now called
`preferred_root_scrollbar_color_scheme`. All the relevant references and
functions have been renamed to reflect this.

Bug: 337904215
Change-Id: I2517747f2a4cef061187daf4191c0fc8660d7fcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5502831
Reviewed-by: Yaroslav Shalivskyy <yshalivskyy@microsoft.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1298816}
NOKEYCHECK=True
GitOrigin-RevId: dbc75481cc094c4daf7337b0dc3d88ab198b57a2