Enable software compositing for threaded web tests.
This change allows software compositing for threaded web tests. I've
run virtual/threaded with this change (and --disable-gpu) and they seem
to "work": no crashes, but a bunch of reference mismatches, which is
expected.

This is desirable for testing some features like ViewTransitions which
originate in blink (and are heavily tested by web tests), but diverge
in implementations in viz for software and gpu compositing.

R=danakj@chromium.org, dtapuska@chromium.org

Change-Id: Ibb3eb36c07134a72c3b9514cf7e805f3eeb3e4bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5527361
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1299217}
NOKEYCHECK=True
GitOrigin-RevId: ba805091c2c36b1792fc4a6698296f0fbbd91c86