RubyLB: Introduce RubyBreakTokenData
It's an optional data of InlineBreakToken.
RubyBreakTokenData and BlockBreakToken are stored together in an
InlineBreakToken as a flexible array.

This CL also remove the declaration of SubBreakTokenAddress(), which
is not defined.

This CL has no behavior changes.

Bug: 324111880
Change-Id: Icd2b226a0441cfa3e4627e2100696d2d699e5e81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5528644
Auto-Submit: Kent Tamura <tkent@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1299279}
NOKEYCHECK=True
GitOrigin-RevId: f99d3226e6a9aedf9b6e4a94f9f0423a30615dea