[a11y] Computes accessible name for svg with image elements.
This change fixes a bug where previously svg
images with a nested title tag were not using the
text from the title as the accessible name, as per
the rules for svg naming defined at:
https://www.w3.org/TR/svg-aam-1.0/

The previous algorithm groupped html images (<img)
and svg images (<image), in  one step to compute
the accessible name by trying to read the alt
attribute. For html, this is correct; However, for
svg images trying to read the alt attribute is
wrong since there is no alt attribute, and the
way to provide accessible descriptions is through
aria.

Test: DumpAccessibilityTreeTest.NameSvgFromContent
AX-Relnotes: n/a.
Bug: 331902964

Change-Id: I0fbd06f4253531b8aaf4bc26f8be9d24df641e7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5515384
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Cr-Commit-Position: refs/heads/main@{#1299329}
NOKEYCHECK=True
GitOrigin-RevId: 4251b33a660778cab0d8364a49b85a7d80bf028d