(Reland) Migrate tfhub_models to gcs first class
This hook is being migrated to be a GCS first class dep. This means
that the sha file is no longer needed. Any updates to this dep will
need to update the gcs dep entry in DEPS, instead of the sha file.

Release builders now have the updated depot_tools version to support
this migration.

Bug: b/337331070
Change-Id: Ib2ef51f525cf08bd07f1dcacb9b221ae97955af3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5529340
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Commit-Queue: Stephanie Kim <kimstephanie@google.com>
Cr-Commit-Position: refs/heads/main@{#1299367}
NOKEYCHECK=True
GitOrigin-RevId: 76ba2cbd5cf8fb721d76d5558307110a23a026de