Remove use_perfetto_client_library build flag in /third_party/blink/renderer/core/timing
This flag was introduced to provide a smooth transition from legacy
tracing macros to Perfetto tracing SDK. Now the transition is complete,
and the flag is unconditionally set to true on all platforms. So we can
remove the flag and the code that is disabled by it.

This CL was uploaded by git cl split.

R=npm@chromium.org

Bug: 42050015, 336718643
Change-Id: I193397184b7485970ebe1d87205d334dd70a8b0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5505352
Commit-Queue: Nicolás Peña <npm@chromium.org>
Auto-Submit: Mikhail Khokhlov <khokhlov@google.com>
Reviewed-by: Nicolás Peña <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1299584}
NOKEYCHECK=True
GitOrigin-RevId: 137125f528f623da0036e04c61b30ee477e6587a